pyproject.toml
[project]
dependencies = [
  "aiofiles>=25.1.0",
  "aiosqlite>=0.22.1",
  "argon2-cffi>=25.1.0",
  "asyncpg>=0.31.0",
  "fastapi[standard-no-fastapi-cloud-cli]>=0.129.0",
  "httpx>=0.28.1",
  "psycopg[binary]>=3.3.3",
  "pydantic-settings>=2.13.1",
  "pyjwt[crypto]>=2.11.0",
  "pytest>=9.0.2",
  "python-multipart>=0.0.22",
  "sqlalchemy[asyncio]>=2.0.46",
  "streamlit>=1.54.0",
  "uvicorn>=0.41.0",
]
description = "This project is a learning resource for FastAPI ecosystem, a modern, fast (high-performance) web framework for building APIs with Python 3.14+ based on standard Python type hints."
name = "fastapi-ecosystem-lab"
readme = "README.md"
requires-python = ">=3.14"
version = "2026.2"

[tool.uv]
package = true

[build-system]
build-backend = "hatchling.build"
requires = ["hatchling"]

[tool.hatch.build.targets.wheel]
packages = ["learn_fastapi"]

[tool.ruff.lint]
extend-ignore = [
  "ARG001", # Ignore "Unused function argument".
  "CPY001", # Ignore "Missing copyright notice".
  "D100", # Ignore "Missing docstring in public module".
  "D101", # Ignore "Missing docstring in public class".
  "D102", # Ignore "Missing docstring in public method".
  "D103", # Ignore "Missing docstring in public function".
  "D104", # Ignore "Missing docstring in public package".
  "T201", # Ignore "print statements".
  "TC001", # Ignore "Move application imports ... into a type-checking block" learn_fastapi types are not compatible with this rule, and it is not critical for Pydantic.
  "TC002", # Ignore "Move third-party library imports ... into a type-checking block" learn_fastapi types are not compatible with this rule, and it is not critical for Pydantic.
  "TC003", # Ignore "Move standard library imports ... into a type-checking block" learn_fastapi types are not compatible with this rule, and it is not critical for Pydantic.
]
extend-select = [] # Here you can add additional rules to be checked by Ruff, if desired.

[tool.ty.rules]
unresolved-import = "ignore"

[dependency-groups]
dev = [
  "pre-commit>=4.0.0",
  "pytest-asyncio>=1.3.0",
  "ruff>=0.15.2",
  "sqlalchemy-stubs>=0.4",
  "sqlalchemy[mypy]>=2.0.47",
  "ty>=0.0.17",
  "types-sqlalchemy>=1.4.53.38",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
addopts = "-v --tb=short"
python_files = ["tests.py", "test_*.py", "*_tests.py"]
testpaths = ["**/tests"]

[tool.ruff.lint.per-file-ignores]
"**/tests/*" = [
  "S101",
] # Ignore "Use of assert detected. The enclosed code will be removed when compiling to optimized bytecode." in test files, as using "assert" is standard practice in tests.
"test_*.py" = [
  "S101",
] # Ignore "Use of assert detected. The enclosed code will be removed when compiling to optimized bytecode." in test files, as using "assert" is standard practice in tests.
"tests/*" = [
  "S101",
] # Ignore "Use of assert detected. The enclosed code will be removed when compiling to optimized bytecode." in test files, as using "assert" is standard practice in tests.
